Implemented automatic line breaking
This is a solution to issue 46 (https://github.com/evilsocket/pwnagotchi/issues/46).

Current implementations of view/voice/state/display objects, leave no room for other place than the update method in view object for text wrapping. It feels wrong, like a dirty hack, but it belongs there in my opinion.
Now, you can write your text, without worrying about it's length (see below about max length), because it automatically breaks in a new line, when the word is too big. Don't worry, whitespace characters will still work if you explicitly want them, but they are not necessary any more.

* max length is 20 characters, it is automatically calculated by subtracting status position from width.
It's already tested with the following text 'I pwn therefore I am.', which is exactly 21 characters long. But, with this implementation it breaks to a new line before 'am.'. You may ask why this happens, since without this implementation, this text fits exactly in one line.
This happens because the character '.' does NOT occupy more than 6 pixels. Any other character (tested) will go out of bounds. Therefore, I concluded that each character occupies 6 pixels. If there is a way to extract that info from the screen driver or the font, please correct me, as that number is hardcoded for now.
In addition, Greek language was tested and it was displayed correctly.

Other:
- added custom text view