Added slack notifications to travis.yml
Should hopefully go to just the travis_builds channel. there was already a secure entry when i got here, not sure what its for but i left it for now until we can verify it is unnecessary. This is the general formatting for this though so we can work from here. 
Thanks!
Signed-off-by: RH <incredincomp@gmail.com>