Enable fstrim.timer to increase SDCard lifetime
fstrim timer triggers weekly to issue TRIM/DISCARD on devices that support it,
which includes SD and MMC devices.

This marks all unused diskspace as available for internal garbage collection
and wear levelling on the card.  It's preferrable to use the timer to the
discard option, as it involves a one-off cost of latency over a global slowing
down of all write or erase requests that the `discard` mount option does.