Make wigle.py work
There were issues with the formatting of the CSV data, and if the original module could successfully post to WiGLE, it would be unnecessarily inflating the WiGLE processing queue with tiny files. A more efficient idea that isn't toxic to the WiGLE community is to collect all CSV lines ready to upload, put them in one "file", and then submit it for processing. The GPS accuracy isn't 0; that would be 0 meters of error, so let's at least increase that to something plausible. Also, removed spaghetti, wasteful assignments, and general waste.