Move system file install into setup.py
The installation of system files has been moved into the 'install' command
instead of running unconditionally.  This makes it possible to create a source code
distribution of that doesn't immediately clobber the build machine's system files.
This will also allow the image build to be simplified as it will only need to copy the
sdist archive instead of multiple directories worth of files.

Signed-off-by: llamasoft <llamasoft@rm-rf.email>