Copy requirements.txt to requirements.in
The idea here is that the requirements.in file should be the loosest allowed
requirements for the project itself, then it should be pip-compile'd into
a concrete requirements.txt that has known working values.  This prevents
unnecessarily pinning packages to specific releases which makes it easier
to keep package versions up-to-date going forward.

Signed-off-by: llamasoft <llamasoft@rm-rf.email>